requirements.txt
argon2-cffi==19.2.0
bcrypt==3.1.7
cffi==1.14.0
dj-database-url==0.5.0
Django==2.2
django-crispy-forms==1.9.0
django-heroku==0.3.1
django-smtp-ssl==1.0
django-taggit==1.2.0
gunicorn==20.0.4
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.2
Paginator==0.5.1
Pillow==7.1.1
psycopg2==2.8.5
pycparser==2.20
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
whitenoise==5.0.1
